add upload-slots-when-seeding
git-svn-id: http://svn.vuze.com/public/client@8228 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/core3/download/DownloadManagerState.java
@@ -57,7 +57,7 @@ DownloadManagerState
 	
 	public static Object[][] PARAMETERS = {
 		{ PARAM_MAX_PEERS,							new Integer( 0 ) },
-		{ PARAM_MAX_UPLOADS_WHEN_SEEDING, 			new Integer( 0 ) },
+		{ PARAM_MAX_UPLOADS_WHEN_SEEDING, 			new Integer( 4 ) },
 		{ PARAM_MAX_UPLOADS_WHEN_SEEDING_ENABLED, 	new Boolean( false ) },
 	};
